% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{get_ncov}
\alias{get_ncov}
\title{Get NCoV data}
\usage{
get_ncov(
  method = c("ncovr", "tidy", "api", "china", "csv"),
  port = c("area", "overall", "news", "rumors"),
  base = "https://lab.isaaclin.cn/nCoV/api/",
  if_en = FALSE
)
}
\arguments{
\item{method}{character.}

\item{port}{character. The port(s). Valid only for the 'api' method}

\item{base}{character. The link of the database api.}
}
\value{
a list
}
\description{
Get NCoV data
}
\examples{
get_ncov('overall')
}
