% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{get_us}
\alias{get_us}
\title{Wrapper for downloading US data}
\usage{
get_us(
  ports = c("states_current", "states_daily", "states_info", "states_screenshots",
    "us_current", "us_daily", "urls")
)
}
\arguments{
\item{ports}{ports}
}
\value{
a list
}
\description{
Wrapper for downloading US data
}
\examples{
us <- get_us("us_current")
us <- get_us("states_daily")
us <- get_us()
}
