% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{plot_foreign_map}
\alias{plot_foreign_map}
\title{Foreign ncov map}
\usage{
plot_foreign_map(
  ncov,
  country,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  bins = c(0, 10, 100, 500, 1000),
  legend_title = "Confirmed",
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  color = "Reds"
)
}
\arguments{
\item{ncov}{ncov data}

\item{country}{country name}

\item{key}{the feature to plot}

\item{bins}{a numberic vector to cut the value to format the value to}

\item{legend_title}{legend title}

\item{legend_position}{legend position}

\item{color}{color palette}
}
\description{
Foreign ncov map
}
