% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{plot_ggmap}
\alias{plot_ggmap}
\title{Plot map with ggplot}
\usage{
plot_ggmap(
  x,
  col_name = "confirmedCount",
  province_language = "chinese",
  show_capitials = TRUE,
  add_title = NA
)
}
\arguments{
\item{x}{data frame with province name and value}

\item{col_name}{column whose value is to display}

\item{province_language}{chinese, english or NA (not shown)}

\item{show_capitials}{logical.}

\item{add_title}{title}
}
\value{
a ggplot map
}
\description{
Plot map with ggplot
}
\examples{
require(ncovr)
library(ggplot2)
library(dplyr)
library(scales)
Sys.setlocale("LC_CTYPE", "Chinese")
ncov <- get_ncov()
ncov$area$date <- as.Date(ncovr:::conv_time(ncov$area$updateTime))

choose_date <- "2020-02-10"
x <- ncov$area[ncov$area$date <= as.Date(choose_date), ]
x <- x[!duplicated(x$provinceName), ]
plot_ggmap(x)
plot_ggmap(x, province_language = "english", show_capitials = TRUE) + 
    annotate(geom = "text", 105, y = 35, label = "ncov2020.org", color = "white", 
        fontface = "bold", size = 10, alpha = 0.5)

library(animation)
ncov_dates <- as.character(rev(unique(ncov$area$date)))
ani.options(interval = 0.5, ani.width = 1000, ani.height = 600)
saveGIF({
    for (i in rev(ncov_dates)) {
        x <- ncov$area[ncov$area$date <= as.Date(i), ]
        x <- x[!duplicated(x$provinceName), ]
        print(plot_ggmap(x, ))
    }
}, movie.name = "map_animation.gif")
}
