% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{plot_map}
\alias{plot_map}
\title{Plot maps for nCoV in China}
\usage{
plot_map(
  x,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  filter = "unkown",
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  method = c("province", "city", "country")
)
}
\arguments{
\item{key}{the column name}

\item{filter}{the rows to remove.}

\item{legend_title}{legend title}

\item{color}{The colors or color function that values will be mapped to, see RColorBrewer::display.brewer.all()}

\item{scale}{"cat" for category, 'log' for numeric}

\item{method}{'province' or 'city'.}

\item{ncov}{the dataset obtained by get_ncov()}
}
\value{

}
\description{
Plot maps for nCoV in China
}
\examples{
ncov <- get_ncov()
plot_map(ncov)
}
