% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series of a country's cases}
\usage{
plot_ts(
  x,
  area,
  area_col = "countryEnglishName",
  date_col = "date",
  ts_col = c("confirmedCount", "curedCount", "deadCount")
)
}
\arguments{
\item{x}{a datafram with "countryEnglishName", "date", "confirmed", "cured", "dead" columns}

\item{area}{name of an selected area}

\item{area_col}{the column name of the area in x}

\item{date_col}{the column name of the date in x}

\item{ts_col}{the column names of the time series in x}
}
\value{
a figure
}
\description{
Plot time series of a country's cases
}
\examples{
ncov_raw <- get_ncov()
ncov_raw$area$date <- as.Date(ncovr:::conv_time(ncov_raw$area$updateTime))
x <- ncov_raw$area[, c('countryEnglishName', 'countryName', 'date', 'confirmedCount', 'curedCount', 'deadCount')] \%>\%
  roup_by(countryEnglishName, date) \%>\%
  summarise(confirmed = sum(confirmedCount),
            cured = sum(curedCount),
            dead = sum(deadCount)) \%>\%
  ungroup()
x <- x[!is.na(x$countryEnglishName) & !x$countryEnglishName == 'China', ]
plot_ts(x, area = 'Korea')
}
