% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{predict_date2}
\alias{predict_date2}
\title{Previous version of predict_date()}
\usage{
predict_date2(
  province,
  ncov = c(ncov, ncovChina),
  ifplot = TRUE,
  addtitle = NA,
  ifchinese = FALSE,
  fontfamily = NA
)
}
\arguments{
\item{province}{province short name}

\item{ncov}{ncov data}

\item{ifplot}{logic.}
}
\value{
a figure.
}
\description{
Previous version of predict_date()
}
\examples{
Sys.setlocale("LC_CTYPE", "Chinese")
ncov <- get_ncov()  #Get the data
ncovChina <- get_ncov("china")
predict_date(province = "jiangsuinchinese", ncov = ncov)
predict_date(province = "meiguoinchinese", ncov = ncov)
predict_date(province = "zhongguoinchinese", ncov = ncovChina)
}
