% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncov-data.R
\name{[.ncov}
\alias{[.ncov}
\title{Subset ncov data}
\usage{
\method{[}{ncov}(x, i, j, latest = TRUE)
}
\arguments{
\item{x}{ncov data, return from `get_ncov()`}

\item{i}{word}
}
\description{
Subset world, china, province, and other countries ncov data
}
