% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\name{get_ncov}
\alias{get_ncov}
\title{Get NCoV data}
\usage{
get_ncov(
  method = c("ncovr", "tidy", "api", "china", "csv", "json"),
  port = c("area", "overall", "news", "rumors"),
  base = "https://lab.isaaclin.cn/nCoV/api/",
  if_en = FALSE
)
}
\arguments{
\item{method}{character.}

\item{port}{character. The port(s). Valid only for the 'api' method}

\item{base}{character. The link of the database api.}
}
\value{
a list
}
\description{
Get NCoV data
}
\examples{
get_ncov("overall")

us_current <- ncovr::get_ncov(method = "api", port = "v1/states/current.json", 
    base = "https://covidtracking.com/api/")

ports <- c("v1/states/current.json", "v1/states/daily.json", 
    "v1/states/info.json", "v1/us/current.json", "v1/us/daily.json", 
    "v1/urls.json", "v1/states/screenshots.json")
us <- ncovr::get_ncov(method = "api", port = ports, 
    base = "https://covidtracking.com/api/")
}
