% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_map.R, R/province_map.R, R/world_map.R
\name{plot_china_map}
\alias{plot_china_map}
\alias{plot_province_map}
\alias{plot_province_map2}
\alias{plot_world_map}
\title{Plot maps for nCoV}
\usage{
plot_china_map(
  ncov,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  filter = "unkown",
  bins = c(0, 10, 100, 500, 1000, 10000),
  legend_title = "Confirmed",
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  color = "Reds"
)

plot_province_map(
  ncov,
  province,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  bins = c(0, 10, 100, 1000),
  map_title = paste0(province, "nCoV")
)

plot_province_map2(
  ncov,
  province,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  bins = c(0, 10, 100, 1000),
  tile_type = NULL,
  map_title = paste0(province, "nCoV")
)

plot_world_map(
  x,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  bins = c(0, 10, 100, 500, 1000, 10000),
  legend_title = "Confirmed",
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  color = "Reds"
)
}
\arguments{
\item{ncov}{ncov data}

\item{key}{the feature to plot}

\item{bins}{a numberic vector to cut the value to format the value to}

\item{legend_title}{legend title}

\item{legend_position}{legend position}

\item{color}{color palette}

\item{province}{which province to plot}

\item{scale}{category or factor}

\item{map_title}{map title
category, only worked while \code{scale} is "cat"}

\item{tile_type}{function to define tile like amap or
\code{\link[leaflet]{addTiles}}, default NULL}
}
\description{
Plot maps for nCoV
}
\details{
For \code{plot_province_map()}, the color is set to white manually if
  there is no ncov.
}
\author{
Yang Cao
}
